+++ Communication/Packets/Incoming/Catalog/GetPromotableRoomsEvent.cs
@@ -16,8 +16,7 @@ internal class GetPromotableRoomsEvent : IPacketEvent
 
     public Task Parse(GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        if (habbo == null)
+        if (session.GetHabbo() is not { } habbo)
             return Task.CompletedTask;
 
         var rooms = _roomFactory.GetRoomsDataByOwnerSortByName(habbo.Id);