+++ Communication/Packets/Incoming/Preferences/SetChatPreferenceEvent.cs
@@ -15,8 +15,7 @@ internal class SetChatPreferenceEvent : IPacketEvent
 
     public async Task Parse(GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        if (habbo == null)
+        if (session.GetHabbo() is not { } habbo)
             return;
 
         var chatPreference = packet.ReadBool();