+++ Communication/Packets/Incoming/Preferences/SetUIFlagsEvent.cs
@@ -8,8 +8,7 @@ internal class SetUIFlagsEvent : IPacketEvent
 {
     public Task Parse(GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        if (habbo == null)
+        if (session.GetHabbo() is not { } habbo)
             return Task.CompletedTask;
 
         habbo.FriendbarState = FriendBarStateUtility.GetEnum(packet.ReadInt());