+++ Communication/Packets/Incoming/Rooms/Action/GiveRoomScoreEvent.cs
@@ -17,10 +17,11 @@ internal class GiveRoomScoreEvent : RoomPacketEvent
 
     public override Task Parse(Room room, GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        if (habbo == null) return Task.CompletedTask;
+        if (session.GetHabbo() is not { } habbo)
+            return Task.CompletedTask;
         if (habbo.RatedRooms.Contains(room.RoomId) || room.CheckRights(session, true))
             return Task.CompletedTask;
+
         var rating = packet.ReadInt();
         switch (rating)
         {
@@ -28,6 +29,7 @@ internal class GiveRoomScoreEvent : RoomPacketEvent
             case 1: room.Score++; break;
             default: return Task.CompletedTask;
         }
+
         using var db = _database.Connection();
         db.Execute("UPDATE `rooms` SET `score` = @score WHERE `id` = @id LIMIT 1", new { score = room.Score, id = room.RoomId });
         habbo.RatedRooms.Add(room.RoomId);