+++ Communication/Packets/Incoming/Rooms/Furni/CreditFurniRedeemEvent.cs
@@ -22,24 +22,30 @@ internal class CreditFurniRedeemEvent : RoomPacketEvent
 
     public override Task Parse(Room room, GameClient session, IIncomingPacket packet)
     {
-        var habbo = session.GetHabbo();
-        var furniture = habbo?.Inventory?.Furniture;
-        if (habbo == null || furniture == null) return Task.CompletedTask;
-        if (!room.CheckRights(session, true)) return Task.CompletedTask;
+        if (session.GetHabbo() is not { Inventory.Furniture: { } furniture } habbo)
+            return Task.CompletedTask;
+        if (!room.CheckRights(session, true))
+            return Task.CompletedTask;
+
         if (_settingsManager.TryGetValue("room.item.exchangeables.enabled") != "1")
         {
             session.SendNotification("The hotel managers have temporarilly disabled exchanging!");
             return Task.CompletedTask;
         }
+
         var exchange = room.GetRoomItemHandler().GetItem(packet.ReadUInt());
-        if (exchange == null) return Task.CompletedTask;
-        if (exchange.Definition.InteractionType != InteractionType.Exchange) return Task.CompletedTask;
+        if (exchange == null)
+            return Task.CompletedTask;
+        if (exchange.Definition.InteractionType != InteractionType.Exchange)
+            return Task.CompletedTask;
+
         var value = exchange.Definition.BehaviourData;
         if (value > 0)
         {
             habbo.Credits += value;
             session.Send(new CreditBalanceComposer(habbo.Credits));
         }
+
         using var db = _database.Connection();
         db.Execute("DELETE FROM `items` WHERE `id` = @id LIMIT 1", new { id = exchange.Id });
         session.Send(new FurniListUpdateComposer());