+++ Communication/Packets/Incoming/Rooms/Furni/GetGroupFurniSettingsEvent.cs
@@ -22,15 +22,18 @@ internal class GetGroupFurniSettingsEvent : IPacketEvent
         var habbo = session.GetHabbo();
         if (habbo == null || !habbo.InRoom || !habbo.TryGetCurrentRoom(out var currentRoom))
             return Task.CompletedTask;
+
         var itemId = packet.ReadUInt();
         var groupId = packet.ReadInt();
         var item = currentRoom.GetRoomItemHandler().GetItem(itemId);
         if (item == null)
             return Task.CompletedTask;
+
         if (item.Definition.InteractionType != InteractionType.GuildGate)
             return Task.CompletedTask;
         if (!_groupManager.TryGetGroup(groupId, out var group))
             return Task.CompletedTask;
+
         session.Send(new GroupFurniSettingsComposer(group, itemId, habbo.Id));
         session.Send(new GroupInfoComposer(group, session, _roomFactory));
         return Task.CompletedTask;